Rename repository files from portfolio_company_share_option_ to company_share_portfolio_option_ pattern
- Renamed 4 repository files to follow new naming convention:
  - portfolio_company_share_option_factor_repository.py -> company_share_portfolio_option_factor_repository.py
  - portfolio_company_share_option_delta_factor_repository.py -> company_share_portfolio_option_delta_factor_repository.py
  - portfolio_company_share_option_price_return_factor_repository.py -> company_share_portfolio_option_price_return_factor_repository.py
  - portfolio_company_share_option_price_factor_repository.py -> company_share_portfolio_option_price_factor_repository.py

- Updated class names from PortfolioCompanyShareOption*Repository to CompanySharePortfolioOption*Repository
- Updated import statements to reference new mapper class names
- Updated group names from 'portfolio_company_share_option' to 'company_share_portfolio_option'
